- Change structure to src/modules/App - Set counter and todo list in separate components - Start cleaning up code
