--- src/App.js
@@ -1,58 +1,0 @@
-import * as React from 'react';
-import { uniqueId } from 'lodash';
-
-import Todo from './components/todo/TodoComponent.js';
-import TodoInput from './components/todo-input/TodoInputComponent.js';
-
-export class App extends React.Component {
-
-  state = {
-    todo: '',
-    todos: [ { text: 'Add your first todo' } ]
-  }
-
-  componentDidUpdate = (prevProps, prevState) => {
-    const { todos: prevTodos } = prevState;
-    const { todos } = this.state;
-    if ( prevTodos.length !== todos.length ) {
-      document.querySelector('#counter').innerText = todos.length;
-    }
-  }
-
-  handleChange = event => this.setState({ todo: event.target.value });
-
-  handleClickAdd = () => {
-    const { todo, todos } = this.state;
-    todo && this.setState({ todos: [ ...todos, { text: todo } ] });
-  };
-
-  handleClickDelete = index => {
-    console.log(`Deleting todo number ${index}`);
-    const { todos } = this.state;
-    this.setState({ todos: [
-      ...todos.slice(0, index),
-      ...todos.slice(index + 1)
-    ]});
-  }
-
-  render() {
-    this.state.todos.forEach((todo, index) => {
-      this.state.todos[index] = { ...todo, id: uniqueId() };
-    });
-    const { todo, todos } = this.state;
-    return (
-      <div className="todo-list">
-        <h1>todos</h1>
-        <p><span id="counter">1</span> remaining</p>
-        <div>
-          {
-            todos.length
-              ? todos.map((todo, index) => <Todo key={todo.id} onClickDelete={() => this.handleClickDelete(index)} text={todo.text} />)
-              : 'You\'re all done 🌴'
-          }
-        </div>
-        <TodoInput handleChange={this.handleChange} handleClickAdd={this.handleClickAdd} todo={todo}/>
-      </div>
-    )
-  }
-}
