+++ src/modules/App/components/counter/CounterComponent.js
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+export default class Counter extends React.Component {
+  render () {
+    return (
+        <p><span id="counter">{this.props.todosLength}</span> remaining</p>    )
+  }
+}
