+++ src/modules/App/components/todo-list/todo-list-styles.css
@@ -1,7 +1,3 @@
-body {
-  font-family: sans-serif;
-}
-
 .todo-list {
   border-radius: 3px;
   box-shadow: 0 2px 7px 0 rgba(0, 0, 0, 0.27);
