+++ src/modules/App/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { render } from 'react-dom';
 
 import { App } from './App';
