+++ src/modules/App/styles.css
@@ -1,0 +1,3 @@
+body {
+  font-family: sans-serif;
+}
