+++ tools/webpack/production.config.js
@@ -5,7 +5,7 @@ const DIST_DIR = path.join(__dirname, '..', '..', 'dist');
 const SRC_DIR = path.join(__dirname, '..', '..', 'src');
 
 module.exports = {
-  entry: `${SRC_DIR}/index.js`,
+  entry: `${SRC_DIR}/modules/App//index.js`,
   module: {
     rules: [
       {
